requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.12.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-excel==2.3.0
drf-renderer-xlsx==1.0.0
et-xmlfile==1.1.0
greenlet==1.1.2
msgpack==1.0.3
neovim==0.3.1
openpyxl==3.1.1
Pillow==9.1.1
PyMySQL==1.0.2
pynvim==0.4.3
pytz==2022.1
sqlparse==0.4.2
